Migrate main LoopFollow target to a synchronized folder group
Convert the main LoopFollow/ group from explicit file enumeration to a
PBXFileSystemSynchronizedRootGroup so files are picked up from disk. This
shrinks project.pbxproj from ~2900 to ~880 lines and removes it as a source
of merge conflicts when files are added, moved, or deleted.

Shared LiveActivity/GlucoseConversion sources that also build into the Live
Activity extension are moved to a new Shared/ folder referenced by both
targets. Info.plist is kept out of the copy phase via a membership exception.

Also remove dead code that was on disk but never in the build: a stale
duplicate RestartLiveActivityIntent, four unreferenced orphan files, and a
leftover Core Data model unused since the initial commit.
