--- LoopFollow/Alarm/SnoozeState.swift
@@ -1,9 +1,0 @@
-// LoopFollow
-// SnoozeState.swift
-
-import Foundation
-
-struct SnoozeState: Codable {
-    var isSnoozed: Bool = false
-    var snoozeUntil: Date?
-}
