--- LoopFollow/Storage/Framework/UserData.xcdatamodeld/UserData.xcdatamodel/contents
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<model type="com.apple.IDECoreDataModeler.DataModel" documentVersion="1.0" lastSavedToolsVersion="16119" systemVersion="19E287" minimumToolsVersion="Automatic" sourceLanguage="Swift" userDefinedModelVersionIdentifier="">
-    <entity name="Nightscout" representedClassName="Nightscout" syncable="YES" codeGenerationType="class">
-        <attribute name="url" optional="YES" attributeType="String"/>
-    </entity>
-    <elements>
-        <element name="Nightscout" positionX="-63" positionY="-18" width="128" height="58"/>
-    </elements>
-</model>
