--- LoopFollow/ViewControllers/AppStateViewController.swift
@@ -1,8 +1,0 @@
-// LoopFollow
-// AppStateViewController.swift
-
-import Foundation
-
-class AppStateViewController {
-    var appStateController: AppStateController?
-}
