+++ Shared/GlucoseLiveActivityAttributes.swift
