+++ Shared/GlucoseSnapshot.swift
